+++ src/main/java/com/texnoera/socialmedia/repository/RefreshTokenRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
 
     Optional<RefreshToken> findByToken(String token);
 
-    Optional<RefreshToken> findByUserId(Long userId);
+    Optional<RefreshToken> findByUserId(Integer userId);
 }
